Correction zone Attaquable Changement IdJoueur en Int32 Reduction de la zone Parcourable en fonction de l'energie du joueur Suppression de methodes doublons

+++ src/client/modele/structures/DetailBatiment.hpp
@@ -18,7 +18,7 @@ struct DetailBatiment {
     double bouclierTaux; ///< Le taux de dégats pris par le bouclier
     sf::Int32 visibilite; ///< Le rayon de visibilité
     sf::Int32 rayonAttaque; ///< Le rayon d'attaque
-    sf::Uint16 idJoueur; ///< L'id du joueur
+    sf::Int32 idJoueur; ///< L'id du joueur
     sf::Int32 prixEnergie; ///< Le cout en energie
     sf::Int32 prixMateriaux; ///< Le cout en materiaux
 };
